Enemy systems refactoring to use jobs
PlayerFollowJob was implemented instead of a previous version of player following system where were 2 foreach cycles to get ref to a player position. and to make an enemy follow the playerPos coordinates. Now it is done using job that needs only a playerEntity.position reference through a SystemAPI.GetComp function, while all the calculations and enemy movements are dealt with inside the linear memory space